Upload izmjene na licensesControlleru

+++ controllers/licensesControllers.js
@@ -61,6 +61,11 @@ const createLicense = async (req, res) => {
 
 const deleteLicense = async (req, res) => {
   const { licenseKey } = req.params;
+  const { encryptedMachineId } = req.body;
+
+  if (!encryptedMachineId) {
+    return res.status(400).json({ message: "Missing encrypted machine ID" });
+  }
 
   try {
     const deletedLicense = await License.findOneAndDelete({
